Rename files and structs for consistent naming conventions
- Rename git_profile_dir.rs to git_profile_config_dir.rs to match struct name
- Rename Git2Config struct to GitConfigGit2 for consistent naming pattern
- Update all imports and references throughout codebase
- Follow naming convention: {Purpose}{Implementation} for structs and files
