+++ src/profile.rs
@@ -1,5 +1,5 @@
 pub mod error;
 pub mod git_config_git2;
-pub mod git_profile_dir;
+pub mod git_profile_config_dir;
 pub mod list;
 pub mod switch;
